+++ Src/AwesomeAssertions/ExceptionAssertionsExtensions.cs
@@ -207,5 +207,40 @@ public static class ExceptionAssertionsExtensions
         return (await task).WithParameterName(paramName, because, becauseArgs);
     }
 
+    /// <summary>
+    /// Asserts that the thrown exception has a parameter which name matches <paramref name="paramName" />.
+    /// </summary>
+    /// <param name="assertions">The <see cref="ExceptionAssertionsTask{TException}"/> containing the thrown exception.</param>
+    /// <param name="paramName">The expected name of the parameter</param>
+    /// <param name="because">
+    /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+    /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+    /// </param>
+    /// <param name="becauseArgs">
+    /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
+    /// </param>
+    public static ExceptionAssertionsTask<TException> WithParameterName<TException>(
+        this ExceptionAssertionsTask<TException> assertions,
+        string paramName,
+        [StringSyntax("CompositeFormat")] string because = "",
+        params object[] becauseArgs)
+        where TException : ArgumentException
+    {
+        Guard.ThrowIfArgumentIsNull(assertions);
+
+        return new ExceptionAssertionsTask<TException>(
+            WithParameterNameAsync(assertions.AsTask(), paramName, because, becauseArgs));
+    }
+
+    private static async Task<ExceptionAssertions<TException>> WithParameterNameAsync<TException>(
+        Task<ExceptionAssertions<TException>> task,
+        string paramName,
+        string because,
+        object[] becauseArgs)
+        where TException : ArgumentException
+    {
+        return (await task).WithParameterName(paramName, because, becauseArgs);
+    }
+
 #pragma warning restore AV1755
 }