+++ Tests/AwesomeAssertions.Specs/Exceptions/AsyncFunctionExceptionAssertionSpecs.cs
@@ -4,6 +4,7 @@ using System;
 using System.Threading.Tasks;
 using AwesomeAssertions.Execution;
 using AwesomeAssertions.Extensions;
+using AwesomeAssertions.Specialized;
 #if NETFRAMEWORK
 using AwesomeAssertions.Specs.Common;
 #endif
@@ -1356,6 +1357,225 @@ public class AsyncFunctionExceptionAssertionSpecs
     }
 
     #endregion
+
+    public class WithInnerException
+    {
+        [Fact]
+        public async Task Can_be_used_with_a_single_type_parameter()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>();
+        }
+
+        [Fact]
+        public async Task Can_be_used_with_a_single_type_parameter_after_an_exact_throw()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            await subject.Should().ThrowExactlyAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>();
+        }
+
+        [Fact]
+        public async Task Can_be_used_with_a_single_type_parameter_after_a_throw_within()
+        {
+            Func<Task> subject = () => Throw.Async(new InvalidOperationException("outer", new ArgumentException("inner")));
+
+            await subject.Should().ThrowWithinAsync<InvalidOperationException>(1.Seconds())
+                .WithInnerException<ArgumentException>();
+        }
+
+        [Fact]
+        public async Task Accepts_a_derived_inner_exception()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException()));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<ArgumentException>();
+        }
+
+        [Fact]
+        public async Task Fails_for_a_different_inner_exception()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentException()));
+
+            Func<Task> act = () => subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>();
+
+            await act.Should().ThrowAsync<XunitException>()
+                .WithMessage("*InvalidOperation*Argument*");
+        }
+
+        [Fact]
+        public async Task Fails_for_a_different_inner_exception_with_a_reason()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentException()));
+
+            Func<Task> act = () => subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>("we want to test the {0} message", "failure");
+
+            await act.Should().ThrowAsync<XunitException>()
+                .WithMessage("*InvalidOperation*because we want to test the failure message*Argument*");
+        }
+
+        [Fact]
+        public async Task Keeps_the_type_of_the_inner_exception_for_further_assertions()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException("someParameter")));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<ArgumentNullException>()
+                .WithParameterName("someParameter");
+        }
+
+        [Fact]
+        public async Task Exposes_the_inner_exception_as_the_awaited_result()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException("someParameter")));
+
+            ExceptionAssertions<ArgumentNullException> assertions =
+                await subject.Should().ThrowAsync<AggregateException>()
+                    .WithInnerException<ArgumentNullException>();
+
+            assertions.Which.ParamName.Should().Be("someParameter");
+        }
+
+        [Fact]
+        public async Task Can_be_chained_for_nested_inner_exceptions()
+        {
+            Func<Task> subject = () => Throw.Async(
+                new AggregateException(new InvalidOperationException("outer", new ArgumentException("inner"))));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>()
+                .WithInnerException<ArgumentException>()
+                .WithMessage("inner");
+        }
+
+        [Fact]
+        public async Task Still_supports_naming_the_thrown_exception_type_explicitly()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerException<AggregateException, InvalidOperationException>();
+        }
+    }
+
+    public class WithInnerExceptionExactly
+    {
+        [Fact]
+        public async Task Can_be_used_with_a_single_type_parameter()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentException()));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerExceptionExactly<ArgumentException>();
+        }
+
+        [Fact]
+        public async Task Rejects_a_derived_inner_exception()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException()));
+
+            Func<Task> act = () => subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerExceptionExactly<ArgumentException>();
+
+            await act.Should().ThrowAsync<XunitException>()
+                .WithMessage("*ArgumentException*ArgumentNullException*");
+        }
+
+        [Fact]
+        public async Task Fails_for_a_different_inner_exception_with_a_reason()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException()));
+
+            Func<Task> act = () => subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerExceptionExactly<ArgumentException>("we want to test the {0} message", "failure");
+
+            await act.Should().ThrowAsync<XunitException>()
+                .WithMessage("*ArgumentException*because we want to test the failure message*ArgumentNullException*");
+        }
+
+        [Fact]
+        public async Task Keeps_the_type_of_the_inner_exception_for_further_assertions()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentNullException("someParameter")));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerExceptionExactly<ArgumentNullException>()
+                .WithParameterName("someParameter");
+        }
+
+        [Fact]
+        public async Task Still_supports_naming_the_thrown_exception_type_explicitly()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new ArgumentException()));
+
+            await subject.Should().ThrowAsync<AggregateException>()
+                .WithInnerExceptionExactly<AggregateException, ArgumentException>();
+        }
+    }
+
+    public class AwaitableBehavior
+    {
+        [Fact]
+        public async Task Can_be_assigned_to_the_task_it_wraps()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            Task<ExceptionAssertions<AggregateException>> assertions = subject.Should().ThrowAsync<AggregateException>();
+
+            (await assertions).Which.InnerException.Should().BeOfType<InvalidOperationException>();
+        }
+
+        [Fact]
+        public async Task Can_be_converted_into_the_task_it_wraps()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            Task<ExceptionAssertions<InvalidOperationException>> assertions = subject.Should()
+                .ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>()
+                .AsTask();
+
+            (await assertions).Which.Should().BeOfType<InvalidOperationException>();
+        }
+
+        [Fact]
+        public async Task Supports_configuring_the_continuation()
+        {
+            Func<Task> subject = () => Throw.Async(new AggregateException(new InvalidOperationException()));
+
+            ExceptionAssertions<InvalidOperationException> assertions = await subject.Should()
+                .ThrowAsync<AggregateException>()
+                .WithInnerException<InvalidOperationException>()
+                .ConfigureAwait(false);
+
+            assertions.Which.Should().BeOfType<InvalidOperationException>();
+        }
+
+        [Fact]
+        public async Task Can_be_used_where_a_task_is_expected()
+        {
+            Func<Task> subject = () => Throw.Async(new InvalidOperationException());
+
+            await Task.WhenAll(
+                subject.Should().ThrowAsync<InvalidOperationException>(),
+                subject.Should().ThrowAsync<Exception>());
+        }
+
+        [Fact]
+        public void Guards_against_a_missing_task()
+        {
+            Action act = () => _ = new ExceptionAssertionsTask<Exception>(null);
+
+            act.Should().Throw<ArgumentNullException>()
+                .WithParameterName("task");
+        }
+    }
 }
 
 internal class AsyncClass